♻️ Refactored UsuarioService to improve "Single Responsability"
