+++ src/main/java/tech/ada/bootcamp/arquitetura/cartaoservice/services/DependenteService.java
@@ -1,0 +1,50 @@
+package tech.ada.bootcamp.arquitetura.cartaoservice.services;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+import tech.ada.bootcamp.arquitetura.cartaoservice.entities.Cartao;
+import tech.ada.bootcamp.arquitetura.cartaoservice.entities.Dependente;
+import tech.ada.bootcamp.arquitetura.cartaoservice.entities.Usuario;
+import tech.ada.bootcamp.arquitetura.cartaoservice.exception.DependenteExistenteException;
+import tech.ada.bootcamp.arquitetura.cartaoservice.payloads.TipoCartao;
+import tech.ada.bootcamp.arquitetura.cartaoservice.payloads.request.CadastroUsuarioRequest;
+import tech.ada.bootcamp.arquitetura.cartaoservice.payloads.request.DependenteRequest;
+import tech.ada.bootcamp.arquitetura.cartaoservice.repositories.DependenteRepository;
+
+@Service
+@RequiredArgsConstructor
+public class DependenteService {
+    
+    private final CriarNovoCartaoService cartaoService;
+    private final DependenteRepository dependenteRepository;
+
+    public List<Dependente> criarDependentes(CadastroUsuarioRequest cadastroUsuarioRequest, Usuario usuario) throws DependenteExistenteException{
+    
+        List<Dependente> dependentes = new ArrayList<Dependente>();
+        for(DependenteRequest dependenteRequest : cadastroUsuarioRequest.getDependentes()){
+            validarDependente(dependenteRequest);
+            Dependente dependente = criarDependente(dependenteRequest, usuario, cadastroUsuarioRequest.getTipoCartao());
+            dependentes.add(dependente);
+        }
+
+        return dependentes;
+    }
+
+    private Dependente criarDependente(DependenteRequest dependenteRequest, Usuario usuario,
+            TipoCartao tipoCartao) {
+        Cartao cartaoAdicional = cartaoService.cadastrarAdicional(dependenteRequest.nome(), tipoCartao, usuario);
+        return new Dependente(null, dependenteRequest.cpf(), dependenteRequest.nome(), usuario, cartaoAdicional, LocalDateTime.now());
+    }
+
+    private void validarDependente(DependenteRequest dependenteRequest) throws DependenteExistenteException {
+        //TODO como testar dependentes salvos durante a transaction?
+        if (dependenteRepository.existsByCpf(dependenteRequest.cpf())) {
+            throw new DependenteExistenteException(dependenteRequest.cpf());
+        }
+    }
+}
